api/requirements.txt
fastapi
mangum
pandas
numpy
gspread
google-auth
python-dotenv
